Remove math utils for `OffsetDateTime`
