+++ time/src/main/java/io/spine/time/OffsetTimes.java
@@ -100,109 +100,6 @@ public final class OffsetTimes {
     }
 
     /**
-     * Obtains new offset time with the specified number of hours added.
-     *
-     * @param value      the value to update
-     * @param hoursToAdd a positive number of hours to add
-     */
-    public static OffsetTime addHours(OffsetTime value, int hoursToAdd) {
-        checkNotNull(value);
-        checkPositive(hoursToAdd);
-        java.time.OffsetTime updated = toJavaTime(value).plusHours(hoursToAdd);
-        return of(updated);
-    }
-
-    /**
-     * Obtains new offset time with the specified number of minutes added.
-     *
-     * @param value        the value to update
-     * @param minutesToAdd a positive number of minutes to add
-     */
-    public static OffsetTime addMinutes(OffsetTime value, int minutesToAdd) {
-        checkNotNull(value);
-        java.time.OffsetTime updated = toJavaTime(value).plusMinutes(minutesToAdd);
-        return of(updated);
-    }
-
-    /**
-     * Obtains new offset time with the specified number of seconds added.
-     *
-     * @param value        the value to update
-     * @param secondsToAdd a positive number of seconds to add
-     */
-    public static OffsetTime addSeconds(OffsetTime value, int secondsToAdd) {
-        checkNotNull(value);
-        checkPositive(secondsToAdd);
-        java.time.OffsetTime updated = toJavaTime(value).plusSeconds(secondsToAdd);
-        return of(updated);
-    }
-
-    /**
-     * Obtains new offset time with the specified number of nanoseconds added.
-     *
-     * @param value      the value to update
-     * @param nanosToAdd a positive number of nanoseconds to add
-     */
-    public static OffsetTime addNanos(OffsetTime value, int nanosToAdd) {
-        checkNotNull(value);
-        checkPositive(nanosToAdd);
-        java.time.OffsetTime updated = toJavaTime(value).plusNanos(nanosToAdd);
-        return of(updated);
-    }
-
-    /**
-     * Obtains new offset time with the specified number of hours subtracted.
-     *
-     * @param value           the value to update
-     * @param hoursToSubtract a positive number of hours to subtract
-     */
-    public static OffsetTime subtractHours(OffsetTime value, int hoursToSubtract) {
-        checkNotNull(value);
-        checkPositive(hoursToSubtract);
-        java.time.OffsetTime updated = toJavaTime(value).minusHours(hoursToSubtract);
-        return of(updated);
-    }
-
-    /**
-     * Obtains new offset time with the specified number of minutes subtracted.
-     *
-     * @param value             the value to update
-     * @param minutesToSubtract a positive number of minutes to subtract
-     */
-    public static OffsetTime subtractMinutes(OffsetTime value, int minutesToSubtract) {
-        checkNotNull(value);
-        checkPositive(minutesToSubtract);
-        java.time.OffsetTime updated = toJavaTime(value).minusMinutes(minutesToSubtract);
-        return of(updated);
-    }
-
-    /**
-     * Obtains new offset time with the specified number of seconds subtracted.
-     *
-     * @param value             the value to update
-     * @param secondsToSubtract a positive number of seconds to subtract
-     */
-    public static OffsetTime subtractSeconds(OffsetTime value, int secondsToSubtract) {
-        checkNotNull(value);
-        checkPositive(secondsToSubtract);
-        java.time.OffsetTime updated = toJavaTime(value).minusSeconds(secondsToSubtract);
-        return of(updated);
-    }
-
-    /**
-     * Obtains new offset time with the specified number of nanoseconds subtracted.
-     *
-     * @param value           the value to update
-     * @param nanosToSubtract a positive number of nanoseconds to subtract
-     */
-    public static OffsetTime subtractNanos(OffsetTime value, int nanosToSubtract) {
-        checkNotNull(value);
-        checkPositive(nanosToSubtract);
-        java.time.OffsetTime updated = toJavaTime(value).minusNanos(nanosToSubtract);
-        return of(updated);
-    }
-
-    /**
      * Returns a ISO 8601 time string corresponding to the passed value.
      */
     public static String toString(OffsetTime value) {
