+++ time/src/test/java/io/spine/time/OffsetTimesShould.java
@@ -21,17 +21,14 @@
 package io.spine.time;
 
 import com.google.common.testing.NullPointerTester;
-import com.google.protobuf.Duration;
 import com.google.protobuf.Timestamp;
-import com.google.protobuf.util.Durations;
-import com.google.protobuf.util.Timestamps;
-import org.junit.Before;
 import org.junit.Test;
+import org.junit.jupiter.api.DisplayName;
 
-import java.text.ParseException;
 import java.util.Calendar;
 
 import static io.spine.base.Time.getCurrentTime;
+import static io.spine.test.DisplayNames.HAVE_PARAMETERLESS_CTOR;
 import static io.spine.test.TestValues.random;
 import static io.spine.test.Tests.assertHasPrivateParameterlessCtor;
 import static io.spine.time.Calendars.at;
@@ -39,19 +36,9 @@ import static io.spine.time.Calendars.getHours;
 import static io.spine.time.Calendars.getMinutes;
 import static io.spine.time.Calendars.getSeconds;
 import static io.spine.time.Calendars.getZoneOffset;
-import static io.spine.time.OffsetTimes.addHours;
-import static io.spine.time.OffsetTimes.addNanos;
-import static io.spine.time.OffsetTimes.addMinutes;
-import static io.spine.time.OffsetTimes.addSeconds;
-import static io.spine.time.OffsetTimes.subtractHours;
-import static io.spine.time.OffsetTimes.subtractNanos;
-import static io.spine.time.OffsetTimes.subtractMinutes;
-import static io.spine.time.OffsetTimes.subtractSeconds;
 import static io.spine.time.EarthTime.HOURS_PER_DAY;
 import static io.spine.time.EarthTime.MINUTES_PER_HOUR;
 import static io.spine.time.EarthTime.SECONDS_PER_MINUTE;
-import static io.spine.time.SiTime.MILLIS_PER_SECOND;
-import static io.spine.time.SiTime.NANOS_PER_MILLISECOND;
 import static io.spine.time.SiTime.NANOS_PER_SECOND;
 import static org.junit.Assert.assertEquals;
 
@@ -61,24 +48,14 @@ import static org.junit.Assert.assertEquals;
  */
 public class OffsetTimesShould extends AbstractZonedTimeTest {
 
-    private Timestamp gmtNow;
-    private OffsetTime now;
-
-    @Override
-    @Before
-    public void setUp() {
-        super.setUp();
-        gmtNow = getCurrentTime();
-        now = OffsetTimes.timeAt(gmtNow, zoneOffset);
-    }
-
     @Test
-    public void have_utility_constructor() {
+    @DisplayName(HAVE_PARAMETERLESS_CTOR)
+    void haveUtilityConstructor() {
         assertHasPrivateParameterlessCtor(OffsetTimes.class);
     }
 
     @Test
-    public void obtain_current() {
+    void obtain_current() {
         final OffsetTime now = OffsetTimes.now(zoneOffset);
         final Calendar cal = at(zoneOffset);
 
@@ -107,113 +84,6 @@ public class OffsetTimesShould extends AbstractZonedTimeTest {
         return LocalTimes.of(hours, minutes, seconds, nanos);
     }
 
-    /*
-     * Math with time
-     */
-
-    @Test
-    public void add_Hour() {
-        final int hoursDelta = random(1, 100);
-        final Duration deltaDuration = Durations2.hours(hoursDelta);
-
-        final Timestamp gmtFuture = Timestamps.add(gmtNow, deltaDuration);
-        final LocalTime expectedFuture = LocalTimes.timeAt(gmtFuture, zoneOffset);
-
-        final LocalTime actualFuture = addHours(now, hoursDelta).getTime();
-
-        assertEquals(expectedFuture, actualFuture);
-    }
-
-    @Test
-    public void subtract_Hour() {
-        final int hoursDelta = random(1, 500);
-        final Duration deltaDuration = Durations2.hours(hoursDelta);
-
-        final Timestamp gmtPast = Timestamps.subtract(gmtNow, deltaDuration);
-        final LocalTime expectedPast = LocalTimes.timeAt(gmtPast, zoneOffset);
-
-        final LocalTime actualPast = subtractHours(now, hoursDelta).getTime();
-
-        assertEquals(expectedPast, actualPast);
-    }
-
-    @Test
-    public void add_minutes() {
-        final int minutesDelta = random(1, 300);
-        final Duration deltaDuration = Durations2.minutes(minutesDelta);
-
-        final Timestamp gmtFuture = Timestamps.add(gmtNow, deltaDuration);
-        final LocalTime expectedFuture = LocalTimes.timeAt(gmtFuture, zoneOffset);
-
-        final LocalTime actualFuture = addMinutes(now, minutesDelta).getTime();
-
-        assertEquals(expectedFuture, actualFuture);
-    }
-
-    @Test
-    public void subtract_minutes() {
-        final int minutesDelta = random(1, 1024);
-        final Duration deltaDuration = Durations2.minutes(minutesDelta);
-
-        final Timestamp gmtPast = Timestamps.subtract(gmtNow, deltaDuration);
-        final LocalTime expectedPast = LocalTimes.timeAt(gmtPast, zoneOffset);
-
-        final LocalTime actualPast = subtractMinutes(now, minutesDelta).getTime();
-
-        assertEquals(expectedPast, actualPast);
-    }
-
-    @Test
-    public void add_seconds() {
-        final int secondsDelta = random(1, 300);
-        final Duration deltaDuration = Durations2.seconds(secondsDelta);
-
-        final Timestamp gmtFuture = Timestamps.add(gmtNow, deltaDuration);
-        final LocalTime expectedFuture = LocalTimes.timeAt(gmtFuture, zoneOffset);
-
-        final LocalTime actualFuture = addSeconds(now, secondsDelta).getTime();
-
-        assertEquals(expectedFuture, actualFuture);
-    }
-
-    @Test
-    public void subtract_seconds() {
-        final int secondsDelta = random(1, 1024);
-        final Duration deltaDuration = Durations2.seconds(secondsDelta);
-
-        final Timestamp gmtPast = Timestamps.subtract(gmtNow, deltaDuration);
-        final LocalTime expectedPast = LocalTimes.timeAt(gmtPast, zoneOffset);
-
-        final LocalTime actualPast = subtractSeconds(now, secondsDelta).getTime();
-
-        assertEquals(expectedPast, actualPast);
-    }
-
-    @Test
-    public void add_millis() {
-        final int millisDelta = random(1, 100_000_000);
-        final Duration deltaDuration = Durations.fromMillis(millisDelta);
-
-        final Timestamp gmtFuture = Timestamps.add(gmtNow, deltaDuration);
-        final LocalTime expectedFuture = LocalTimes.timeAt(gmtFuture, zoneOffset);
-
-        final LocalTime actualFuture = addNanos(now, millisDelta).getTime();
-
-        assertEquals(expectedFuture, actualFuture);
-    }
-
-    @Test
-    public void subtract_millis() {
-        final int millisDelta = random(1, 999_999);
-        final Duration deltaDuration = Durations.fromMillis(millisDelta);
-
-        final Timestamp gmtPast = Timestamps.subtract(gmtNow, deltaDuration);
-        final LocalTime expectedPast = LocalTimes.timeAt(gmtPast, zoneOffset);
-
-        final LocalTime actualPast = subtractNanos(now, millisDelta).getTime();
-
-        assertEquals(expectedPast, actualPast);
-    }
 
     @Test
     public void pass_null_tolerance_test() {
@@ -225,111 +95,11 @@ public class OffsetTimesShould extends AbstractZonedTimeTest {
                 .testAllPublicStaticMethods(OffsetTimes.class);
     }
 
-    /*
-     * Illegal args. check for math with hours.
-     */
-    
-    @Test(expected = IllegalArgumentException.class)
-    public void not_accept_negative_hours_to_add() {
-        addHours(now, -5);
-    }
-
-    @Test(expected = IllegalArgumentException.class)
-    public void not_accept_zero_hours_to_add() {
-        addHours(now, 0);
-    }
-
-    @Test(expected = IllegalArgumentException.class)
-    public void not_accept_negative_hours_to_subtract() {
-        subtractHours(now, -6);
-    }
-
-    @Test(expected = IllegalArgumentException.class)
-    public void not_accept_zero_hours_to_subtract() {
-        subtractHours(now, 0);
-    }
-
-    /*
-     * Illegal args. check for math with minutes.
-     */
-
-    @Test(expected = IllegalArgumentException.class)
-    public void not_accept_negative_minutes_to_add() {
-        addMinutes(now, -7);
-    }
-
-    @Test(expected = IllegalArgumentException.class)
-    public void not_accept_zero_minutes_to_add() {
-        addMinutes(now, 0);
-    }
-
-    @Test(expected = IllegalArgumentException.class)
-    public void not_accept_negative_minutes_to_subtract() {
-        subtractMinutes(now, -8);
-    }
-
-    @Test(expected = IllegalArgumentException.class)
-    public void not_accept_zero_minutes_to_subtract() {
-        subtractMinutes(now, 0);
-    }
-
-    /*
-     * Illegal args. check for math with seconds.
-     */
-
-    @Test(expected = IllegalArgumentException.class)
-    public void not_accept_negative_seconds_to_add() {
-        addSeconds(now, -25);
-    }
-
-    @Test(expected = IllegalArgumentException.class)
-    public void not_accept_zero_seconds_to_add() {
-        addSeconds(now, 0);
-    }
-
-    @Test(expected = IllegalArgumentException.class)
-    public void not_accept_negative_seconds_to_subtract() {
-        subtractSeconds(now, -27);
-    }
-
-    @Test(expected = IllegalArgumentException.class)
-    public void not_accept_zero_seconds_to_subtract() {
-        subtractSeconds(now, 0);
-    }
-
-    /*
-     * Illegal args. check for math with millis.
-     */
-
-    @Test(expected = IllegalArgumentException.class)
-    public void not_accept_negative_millis_to_add() {
-        addNanos(now, -500);
-    }
-
-    @Test(expected = IllegalArgumentException.class)
-    public void not_accept_zero_millis_to_add() {
-        addNanos(now, 0);
-    }
-
-    @Test(expected = IllegalArgumentException.class)
-    public void not_accept_negative_millis_to_subtract() {
-        subtractNanos(now, -270);
-    }
-
-    @Test(expected = IllegalArgumentException.class)
-    public void not_accept_zero_millis_to_subtract() {
-        subtractNanos(now, 0);
-    }
-
-    /*
-     * Stringification
-     */
-    
     @Override
-    protected void assertConversionAt(ZoneOffset zoneOffset) throws ParseException {
-        final OffsetTime now = OffsetTimes.now(zoneOffset);
-        final String str = OffsetTimes.toString(now);
-        final OffsetTime parsed = OffsetTimes.parse(str);
+    protected void assertConversionAt(ZoneOffset zoneOffset) {
+        OffsetTime now = OffsetTimes.now(zoneOffset);
+        String str = OffsetTimes.toString(now);
+        OffsetTime parsed = OffsetTimes.parse(str);
         assertEquals(now, parsed);
     }
 }
